# Conflicts:
#	JWT-with-Spring-Security/src/main/java/com/spring/jwt/controller/PendingBookingController.java

+++ JWT-with-Spring-Security/src/main/java/com/spring/jwt/repository/PendingBookingRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PendingBookingRepository extends JpaRepository<PendingBooking, Long> {
+public interface PendingBookingRepository extends JpaRepository<PendingBooking, Integer> {
 
 }
